requirements.txt
openai==1.76.0
streamlit==1.12.0
azure-cosmos==4.9.0
python-dotenv==1.1.0
